requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.7
django-crispy-forms==2.4
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
redis==7.0.1
reportlab==4.4.4
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
